Added screenshots of developing the home,about and contact page with the header and footer.
Used a flex container on the body of the html styling so the header and
footer will take the full width of the screen and the main will fill
the space between.

Added further styling for the content added to the about and contact pages.
Added content to the html files of the about and contacts pages.
Added images to the html files and the repository.